python-django/pyproject.toml
[tool.poetry]
name = "rover-api"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
packages = [
    { include = "apps" },
    { include = "domains" },
    { include = "infra" },
]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.0.3"
djangorestframework = "^3.13.1"
requests = "^2.27.1"
django-cors-headers = "^3.11.0"
attrs = "^21.4.0"
celery = {extras = ["redis"], version = "^5.2.3"}
Flask = "^2.0.3"
pydantic = "^1.9.0"
"zope.event" = "^4.5.0"

[tool.poetry.dev-dependencies]
flake8 = "^4.0.1"
black = "^22.1.0"
isort = "^5.10.1"
pytest = "^7.0.1"
responses = "^0.19.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
